Also harvest non-Wacans when farming using Wacans
Was previously not doing any harvesting on plots intended for Wacans.
This meant the plots could get clogged up with other plants which
wouldn't get harvested.

This was particularly a problem when there were parasitic plants on the
field, which could spread to the Wacans, then just get left alone to
spread further.