Farm-hand: Filter point quest berries to ones which are quick to grow
When picking berries to plant for a Farm Point quest, will now only pick
ones which take 15 minutes or less to grow.

Separate, but also in this commit because of reasons:
I recently added a feature where the bot will farm another berry if it
is trying to farm one berry, but we ran out of any more to plant.
This now no longer happens with Kasibs and Kebias as switching to other
berries was breaking their unique farming methods.