farm-hand: Fix issue when farming harvestAmount=1 berries
When we only had 1 Custap berry, the task as forgetting that it was
trying to farm a harvestAmount=1 berry because it would try to pivot to
planting other plants while we didn't have any Custaps left to plant.
Modified this to still remember the "true target" berry, and base some
of the farming strategy decisions on that rather than the main target
berry variable, which is sometimes this pivot berry.