Keep mining tasks alive until 4k diamonds
This means the script is kinda happy with the diamond count being
anywhere from 3k to 4k. It'll only start mining for more below 3k, but
will then keep going until it reaches 4k.

This stops a bunch of log spam of the tasks starting and stopping every
few minutes when the diamond hatchery helpers are keeping the count
right around 3k.